Autoconfig Messages
More autoconfig messages.  ikec exits if the path for the configurations is null.  The problem is if an undefined configuration is attempted to be loaded, the user's path is checked first.  If the config is not found (which it won't be), then the global path is checked.  If there is no global path, ikec was throwing a double free error.  Now it exits gracefully.